added conditional disabling tests for env variables
